테스트 후 수정 (#11)
* add: User Entity 생성

* update: User Entity 생성자 추가

* add: SignupRequestDto 생성. Validation 적용

* update: 회원가입 기능 구현(Bcrypt 암호화 적용)

* update: User Entity에 유저를 deactivate 하는 메서드 추가

* update: User 회원탈퇴 기능 추가

* update: User Profile 조회, 수정 기능 추가

* fix: newsfeed 명을 post로 변경

* login feature 수정중6 (isTokenValid 수정중)

* refactor: UserService.java 파일 merge

* refactor: UserService.java 파일 merge

* fix: API 수정

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken 으로 토큰 새로발급받는 기능 구현 완료

* RefreshToken1

* RefreshToken2

* RefreshToken3

* logout 수정중1

* logout 수정중2

* logout 구현 완료

* fix: 메서드명 직관적이게 수정

* fix: 일관성 맞춰서 url 수정

* test: 기능테스트

---------

+++ src/main/java/com/sparta/newsfeedapp/service/CommentService.java
@@ -8,7 +8,7 @@ import com.sparta.newsfeedapp.entity.Post;
 import com.sparta.newsfeedapp.entity.User;
 import com.sparta.newsfeedapp.repository.CommentRepository;
 import com.sparta.newsfeedapp.repository.PostRepository;
-import com.sparta.newsfeedapp.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,9 +19,9 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class CommentService {
-    CommentRepository commentRepository;
-    UserRepository userRepository;
-    PostRepository postRepository;
+
+    private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
 
     public Long findMaxIdOfCommentTable(){
         return Optional.ofNullable(commentRepository.findMaxId()).orElse((long)0);
@@ -36,13 +36,14 @@ public class CommentService {
         return new CommentResponseDto(comment);
     }
 
-//    public List<Comment> getComments(long postId){
-//        return commentRepository.findByPostId(postId);
-//    }
+    @Transactional
+    public ResponseEntity<String> updateComment(CommentUpdateRequestDto requestDto, Long commentId, User user) {
 
-    public ResponseEntity<String> updateComment(CommentUpdateRequestDto requestDto, Long commentId) {
         if (commentRepository.existsById(commentId)) {
             Comment comment = commentRepository.findById(commentId).orElseThrow(NullPointerException::new);
+            if (!comment.getUser().getId().equals(user.getId())){
+                throw new IllegalArgumentException("본인 댓글만 수정할 수 있습니다.");
+            }
             comment.update(requestDto);
             return new ResponseEntity<>("성공적으로 수정했습니다. (" + comment.getContent() + ")", HttpStatus.OK);
         } else {
@@ -50,9 +51,13 @@ public class CommentService {
         }
     }
 
-    public ResponseEntity<String> deleteComment(Long commentId){
+    public ResponseEntity<String> deleteComment(Long commentId, User user){
+
         if (commentRepository.existsById(commentId)) {
             Comment comment = commentRepository.findById(commentId).orElseThrow(NullPointerException::new);
+            if (!comment.getUser().getId().equals(user.getId())){
+                throw new IllegalArgumentException("본인 댓글만 삭제할 수 있습니다.");
+            }
             commentRepository.deleteById(commentId);
             return new ResponseEntity<>("성공적으로 삭제했습니다.", HttpStatus.OK);
         } else {
